// frontend/app/types/tag-types.ts
export enum tagTypes {
    user = 'user',
    product = 'product',
    announcement = 'announcement',
    payment = 'payment',
    order = 'order',
    contact = 'contact',
}


export const tagTypesList = [
    tagTypes.user,
    tagTypes.product,
    tagTypes.announcement,
    tagTypes.payment,
    tagTypes.contact,
];